Changed id headers to result in an id of `h<number>:<id>`
While this does make it harder to query select these elements, that is probably a good thing. Doing this has no real impact on linking to these elements, but it could help prevent issues with conflicting ids. If there is a need to select these headers by id in CSS, just escape the `:`.